[MOVE] : MyPage move #5
