--- app/src/main/java/org/sopt/dosopttemplate/presentation/main/MainEvent.kt
@@ -1,6 +1,0 @@
-package org.sopt.dosopttemplate.presentation.main
-
-
-sealed class MainEvent {
-    data class LogOut(val p: Unit) : MainEvent()
-}
