+++ app/src/main/java/org/sopt/dosopttemplate/presentation/mypage/MyPageEvent.kt
@@ -1,0 +1,6 @@
+package org.sopt.dosopttemplate.presentation.mypage
+
+
+sealed class MyPageEvent {
+    data class LogOut(val p: Unit) : MyPageEvent()
+}
