+++ app/src/main/java/org/sopt/dosopttemplate/presentation/mypage/MyPageViewModel.kt
@@ -1,7 +1,5 @@
-package org.sopt.dosopttemplate.presentation.main
+package org.sopt.dosopttemplate.presentation.mypage
 
-import androidx.lifecycle.LiveData
-import androidx.lifecycle.MutableLiveData
 import androidx.lifecycle.ViewModel
 import androidx.lifecycle.viewModelScope
 import dagger.hilt.android.lifecycle.HiltViewModel
@@ -9,17 +7,14 @@ import kotlinx.coroutines.flow.MutableSharedFlow
 import kotlinx.coroutines.flow.asSharedFlow
 import kotlinx.coroutines.launch
 import org.sopt.dosopttemplate.domain.repo.AuthRepo
-import org.sopt.dosopttemplate.presentation.model.User
-import org.sopt.dosopttemplate.presentation.signin.SignEvent
-import org.sopt.dosopttemplate.presentation.signin.SignInViewModel
 import javax.inject.Inject
 
 @HiltViewModel
-class MainViewModel @Inject constructor(
+class MyPageViewModel @Inject constructor(
     private val authRepo: AuthRepo
 ) : ViewModel() {
 
-    private val _eventFlow = MutableSharedFlow<MainEvent>()
+    private val _eventFlow = MutableSharedFlow<MyPageEvent>()
     val eventFlow = _eventFlow.asSharedFlow()
 
     fun signOut() {
@@ -29,10 +24,10 @@ class MainViewModel @Inject constructor(
     fun getUser() = authRepo.getUser()
 
     fun logoutEvent() {
-        event(MainEvent.LogOut(Unit))
+        event(MyPageEvent.LogOut(Unit))
     }
 
-    private fun event(event: MainEvent) {
+    private fun event(event: MyPageEvent) {
         viewModelScope.launch { _eventFlow.emit(event) }
     }
 }
